pyproject.toml
[project]
name = "siecom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.1",
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-phonenumber-field[phonenumberslite]>=8.1.0",
    "django-tree-queries>=0.21.1",
    "graphene-django>=3.2.3",
    "psycopg2-binary>=2.9.10",
    "pytest-cov>=7.0.0",
    "python-decouple>=3.8",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "ruff>=0.13.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest-django>=4.11.1",
    "pytest-xdist>=3.8.0",
]
